+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 * Image class: shape and sizes attributes are populated in the constructor.
+* Z-projection module: change output file naming (do not create a zprojection/ sub-folder, add _vPR<projection> suffix, use same basename as projected file for log file).
 
 ### Removed